requirements.txt
# Core dependencies
requests>=2.31.0
matplotlib>=3.5.0

# For environment variable loading
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Data handling
pandas>=1.5.0
numpy>=1.21.0
types-requests